refactor(core): Improve DX in the MenuBar API to make it more intuitive and readable
Replace the old `config` method, which was verbose and required separate lists for properties, shortcuts, and signals.

Introduce a new `configure` method that uses a single `controls` dictionary. This allows all configuration for an action (properties and signals) to be grouped into a single block, making it much easier to read and maintain.